requirements.txt
Django
PIL
django-piston
restkit
wsgiref
simplejson
djangorestframework
django-filter
